Finished carrying. Added sliding. Cleaned up climbing.
Added the ability to use an empty GameObject named "CarryingPoint" to
use as the point the player holds when carrying an object. If the
"CarryingPoint" object is not present, the item will be held by the
origin and a warning will log to the console.
All carryable objects should be on the "Carryable" layer.
The "Carryable" layer does not collide with the "Player" layer.
Tagged Em and Rawl as "Player" and moved them to the "Player" layer.
Cleaned up carrying code.

Added some basic functionality for the slide ability.

The climbing script now assumes that it is on the same object as the
player movement script.
The climbing script now moves the player a set distance from the wall as
well as constraining the player's rotation to line up with the
collider's z axis.